Theas.js changes:
TH.get now searches for controls only in the Theas form.

Added convenience method getval() that can be used instead of get().val() ... which simplifies coding when the call to get may not return a control.  Added corresponding setval() method for consistency.

Added parameter namPrefix to decodeAll() to support URL-decoding form fields that are not persisted as Theas params.

Changed receiveAsync() and  receiveAsyncError() to display a modal message instead of navigating to the home page if there is an error.

Added parameter performUpdate to submitForm() to simplify initiating a POST that isn't intended to perform a server-side update.

Added method raiseError() to set the th:ErrorMessage and display the modal.

haveError() can now parse pipe-delimited error messsages to set the modal title.

Various modal improvements.